On EFI, GRUB should look for its minimal grub.cfg in /boot/efi/EFI/rosa
...rather than in /boot/efi/EFI/rosa/grub2-efi.

/boot/efi/EFI/rosa/grub2-efi/ is where the signed GRUB binaries are
placed by grub2-efi RPM package.

When grub2-install runs in the installed system, it takes the appropriate
image from /boot/efi/EFI/rosa/grub2-efi/ and copies it to
/boot/efi/EFI/rosa/ with a proper name. It also generates a minimal grub.cfg
in /boot/efi/EFI/rosa/ that instruct grub to use /boot/grub2/grub.cfg,
which can be updated separately (by update-grub2 or grub2-*mkconfig).

The problem was, grub.efi looked for that minimal grub.cfg in
/boot/efi/EFI/rosa/grub2-efi/, because that path  was set as a prefix when
grub.efi was created. As a result, the system failed to boot.

Should be fixed now.