Install the non-EFI bootloader to the boot disk on non-EFI systems only
And do it in %posttrans, to make sure the config files, the files from
/etc/grub.d/, etc., have finally settled at their final locations.

grub2 package can be installed either standalone (on the systems with
legacy BIOS) and as a dependency for grub2-efi* (on UEFI systems).
In the latter case, that package is responsible for the installation
of the boot loader and for updating of /boot/grub2/grub.cfg.
grub2 package should not do it.