grub2-efi-common should require efibootmgr and grub2 packages
efibootmgr can be used to view boot configuration on UEFI systems as
well as change it (via grub2-install or directly).

grub2 package provides the useful tools and convenience scripts like
update-grub2, the configuration scripts from /etc/grub.d, etc., which
are used on UEFI systems as well. So, it is now required too.