Rename grub2-efi-common back into grub2-efi
grub2-efi will contain the unsigned GRUB binaries, while the signed ones
will be kept in grub2-efi-signed, in a separate project.

It can be beneficial to keep grub2-efi-signed separate, for example,
to be able to fix the GRUB tools and update grub2-efi RPM without having
to sign the EFI images again.

All packages from a project are either published or declined as a whole.
So one cannot rebuild GRUB to fix /etc/grub.d/10_linux or grub2-mkconfig
and then publish only the packages containing these fixes. If the signed
GRUB binaries were kept in the same project (even in a different RPM),
one would have to update them as well. And that means signing them
again, etc, which is not always possible (token expiration, org issues,
and so on).

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>