grub2-efi: require grub2-efi-signed >= 2.02-8
The signed grub binaries are in a separate project, grub2-efi-signed,
for now. It can be beneficial to keep it separate, for example, to be able
to fix the GRUB tools or the scripts from /etc/grub.d/ and update
grub2-efi RPM without having to sign the EFI images again.

This is why grub2-efi requires version >= some_minimal_version of
grub2-efi-signed rather than the same EVRD that it has.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>