Use unifont v11.0.02 in PCF format rather than v6.2 in TTF format
I found that GRUB 2.02 failed to load the OS on some 32-bit systems
(both on a desktop and on a EeePC 1012CE laptop) if the user does not
hit any button until the boot timeout expires. The system either hung
or, most of the time, rebooted automatically after GRUB had printed smth
like "Loading Linux ...".

If the user hit Enter before the timeout had expired, the system booted
normally.

The problem did not show up on 64-bit systems.

I bisected our patchset and found that unifont was somehow involved: if
the build system of GRUB found unifont.ttf, the problem was likely to
happen. If the font was not found, the boot process went OK.

I guess that it might be related to processing of TrueType fonts:
neither Fedora/RHEL nor Ubuntu use these for GRUB.

Similar to Fedora, I have updated the build instructions, so that a PCF
variant of the font is used. Let us use the most recent version of it,
just in case.

Seems to fix the issue.

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>