linuxefi: disallow unsigned kernels again
Ubuntu restored the patch to disable linuxefi fallback for unsigned
kernels if SB was enabled. Let us restore it as well.

If they can manage this way now, ROSA can surely too.

Of course, this should be thoroughly tested for both SB and non-SB
cases, with the common builds and the builds with enhanced security.